% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audio_features.R
\docType{data}
\name{audio_features}
\alias{audio_features}
\title{Audio Features of Linkin Park Songs}
\format{A tibble of 6577 observations of 39 variables

See the Spotify API documentation for audio features.  https://developer.spotify.com/documentation/web-api/}
\usage{
audio_features
}
\description{
A collection of spotify audio features of Linkin Park Albums
}
\keyword{datasets}
