% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/billboard_albums.R
\docType{data}
\name{billboard_albums}
\alias{billboard_albums}
\title{Peak Position of Linkin Park Albums on Billboard}
\format{A tibble with 10 observations of 4 variables
\itemize{
\item \code{album_title}: Name of the album charted
\item \code{peak_position}: The highest charted position
\item \code{date_peaked}: Date album peaked on the chart (Format YYYY-MM-DD)
\item \code{weeks_on_chart}: Number of weeks the album was charted.
}}
\usage{
billboard_albums
}
\description{
A data set containing peak positions of charted Linkin Park Albums
from the Billboard Hard Rock Chart. This data was scraped from billboard.com
}
\keyword{datasets}
