% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/billboard_albums.R
\docType{data}
\name{billboard_songs}
\alias{billboard_songs}
\title{Peak Position of Linkin Park Songs on Billboard}
\format{A tibble with 10 observations of 4 variables
\itemize{
\item \code{song_title}: Name of the album charted
\item \code{peak_position}: The highest charted position
\item \code{date_peaked}: Date album peaked on the chart (Format YYYY-MM-DD)
\item \code{weeks_on_chart}: Number of weeks the album was charted.
}}
\usage{
billboard_songs
}
\description{
A data set containing peak positions of charted Linkin Park Songs
from the Billboard Hard Rock streaming songs Chart. This data was scraped from billboard.com
}
\keyword{datasets}
