% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audio_features.R
\docType{data}
\name{genius_lycics}
\alias{genius_lycics}
\title{Lyrics of Linkin Park Songs}
\format{A tibble of 8760 observations of 6 variables
\itemize{
\item \code{artist}: Name of the artist
\item \code{album} : Name of the album
\item \code{track_title}: Name of the track
\item \code{track_number}: Order track appears on album
\item \code{line}: Line number of lyric
\item \code{Lyric}: text of a given line
}}
\usage{
genius_lycics
}
\description{
A collection of genius lyrics of Linkin Park Albums
}
\keyword{datasets}
