% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors_palettes.R
\docType{data}
\name{linkin_park_palettes}
\alias{linkin_park_palettes}
\title{linkin Park Album Inspired Color Palettes}
\format{An object of class \code{list} of length 4.}
\usage{
linkin_park_palettes
}
\description{
linkin Park Album Inspired Color Palettes
}
\keyword{internal}
