% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkinpark-package.R
\docType{package}
\name{linkinpark-package}
\alias{linkinpark-package}
\title{Linkin Park A Data Package About the Band}
\description{
Linkin Park is an American rock band that rose to fame with their debut album \emph{Hybrid Theory},
a new metal album that defined how they would be seen. They followed up their success with \emph{Meteora} which
met with similar success and began to explore more experimental sounds with their third album \emph{Minutes To Midnight}.
They have produced 7 studio albums, 2 remix albums, as well as many other EP, LP, and collaborations.
The band is truly iconic, being the first rock band to get 1 billion views on Youtube, while ranking highly on many other metrics.
Their vision of a Hybrid Sound between Rock and Rap has influenced many artists in many genres.

This package aims to collect publicly available data in a consistent tidy format. Data is scraped from a verity of sources
including: RIAA, Wikipedia, the Linkin Park Wiki, and \href{http://www.reddit.com/r/linkinpark}{/r/Linkinpark}. More data can be found
using the {spotifyr} package that tracks data about the actual audio features.
}
