% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_to_file.R
\name{path_to_file}
\alias{path_to_file}
\title{Get file path to bundled csv files}
\source{
This function is adapted from `readxl::readxl_example()`.
}
\usage{
path_to_file(path = NULL)
}
\arguments{
\item{path}{Name of file in quotes with extension;
`"penguins.csv"` and `"penguins_raw.csv"` will work.
If `NULL`, the example files will be listed.}
}
\description{
linkinpark comes bundled with csv files for included datasets in `inst/extdata`
directory. This is a helper function for easy access.
}
\examples{
path_to_file()
path_to_file("riaa_lp")
head(read.csv(path_to_file("riaa_lp")))
}
