% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reddit_polls.R
\docType{data}
\name{reddit_top_100_songs}
\alias{reddit_top_100_songs}
\title{Top 100 Linkin Park Songs According to Reddit}
\format{A tibble with 103 observations of 4 variables.
\itemize{
\item \code{Rank}: the ranking of the song
\item \code{Track}: the track name
\item \code{Album}: the album name
\item \code{score}: the score (out of 100)
}}
\usage{
reddit_top_100_songs
}
\description{
This data comes from \href{http://www.reddit.com/r/linkinpark}{/r/linkinpark}
From a poll conducted using this \href{http://www.allourideas.org/linkinparksongs}{polling service} where users
selected their favorite song in head to head pairings. There were over 47,000 pairwise votes.
and this was the "official" list created from the results on 2020-07-02.
}
\keyword{datasets}
